Cleanup dir structure for new asset caching
